requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
